attach_render and commit are all that's needed
attach_render tells the output that a "new" buffer has been prepared
(even if we haven't changed it).  We need to call that and then commit
it to keep the render loop going.

Software cursors will freeze momentarily during layout updates, but I
suspect that this is not as easily fixed as it sounds.  You can force
software cursors by running:

    WLR_NO_HARDWARE_CURSORS=1 ./dwl